% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minioR-package.R
\docType{package}
\name{minioR}
\alias{minioR-package}
\alias{minioR}
\title{minioR: A Simple MinIO Client for R}
\description{
`minioR` provides a set of functions to interact with
Amazon S3–compatible object storage services such as MinIO.
It allows uploading, downloading, copying, listing, and
deleting objects directly from R.
}
\details{
The package is designed for data engineering workflows,
automation, and testing scenarios, offering a simple and
explicit interface.
}
\section{Configuration}{

Access to MinIO is configured using environment variables:
\describe{
  \item{MINIO_ENDPOINT}{Endpoint URL (e.g. \code{http://localhost:9000})}
  \item{AWS_SECRET_ACCESS_KEY}{Access key}
  \item{AWS_ACCESS_KEY_ID }{Secret key}
  \item{AWS_SIGNATURE_VERSION}{Use value = 2}
  \item{MINIO_REGION}{Region (optional)}
  \item{MINIO_USE_HTTPS}{Logical value indicating whether to use HTTPS}
}
}

\section{Typical workflow}{

\enumerate{
  \item Configure environment variables
  \item Upload objects using \code{minio_put_object()}
  \item Download objects using \code{minio_download_object()}
}
}

\seealso{
\itemize{
  \item \code{\link{minio_put_object}}
  \item \code{\link{minio_fput_object}}
  \item \code{\link{minio_get_object}}
  \item \code{\link{minio_get_object_metadata}}
  \item \code{\link{minio_get_csv}}
  \item \code{\link{minio_get_parquet}}
  \item \code{\link{minio_download_object}}
  \item \code{\link{minio_copy_object}}
  \item \code{\link{minio_list_objects}}
  \item \code{\link{minio_object_exists}}
  \item \code{\link{minio_remove_object}}
}
}
\author{
\strong{Maintainer}: Javier Ramos \email{jaramosg@ine.gob.cl}

Other contributors:
\itemize{
  \item Victor Ballesteros \email{vaballesteros@ine.gob.cl} [contributor]
  \item Catalina Quijada \email{caquijada@ine.gob.cl} [contributor]
  \item Gabriel Molina \email{gemolinah@ine.gob.cl} [contributor]
}

}
